requirements.txt
asgiref==3.8.1
dj-database-url==3.0.0
Django==5.2.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
pillow==11.2.1
python-decouple==3.8
sqlparse==0.5.3
typing_extensions==4.14.0
whitenoise==6.9.0
